# Micro-optimization for random.choices().  Convert range() to repeat().