Borrowing from the technique used in the *timeit* module, loop with *repeat()* instead of *range()* to avoid unnecessary creation and destruction of integer objects.  Gives a modest speed-up (in the 5% range).

Patched:
```
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, cum_weights=w, k=2)'
200000 loops, best of 11: 1.82 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, cum_weights=w, k=1000)'
1000 loops, best of 11: 388 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, cum_weights=w, k=1000)'
500 loops, best of 11: 390 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, k=2)'
200000 loops, best of 11: 1.43 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' 'choices(p, k=1000)'
1000 loops, best of 11: 255 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' 'choices(p, k=1000)'
1000 loops, best of 11: 255 usec per loop
```

Baseline:
```
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, cum_weights=w, k=2)'
200000 loops, best of 11: 1.95 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, cum_weights=w, k=1000)'
500 loops, best of 11: 398 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, cum_weights=w, k=1000)'
500 loops, best of 11: 402 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, k=2)'
200000 loops, best of 11: 1.5 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, k=1000)'
1000 loops, best of 11: 271 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, k=1000)'
1000 loops, best of 11: 272 usec per loop
```