@rhettinger 2019-02-16
The above timings were on Clang.  The following were run with GCC-8 and show a consistent 5% to 6% improvement.

Patched:

```
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, k=2)'
200000 loops, best of 11: 1.27 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, cum_weights=w, k=2)'
200000 loops, best of 11: 1.68 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, k=1000)'
1000 loops, best of 11: 211 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, cum_weights=w, k=1000)'
1000 loops, best of 11: 363 usec per loop
```

Baseline:

```
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, k=2)'
200000 loops, best of 11: 1.34 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, cum_weights=w, k=2)'
200000 loops, best of 11: 1.79 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, k=1000)'
1000 loops, best of 11: 225 usec per loop
$ ./python.exe -m timeit -r11 -s 'from random import choices' -s'p=range(20)' -s'w=list(map(float, range(20)))' 'choices(p, cum_weights=w, k=1000)'
1000 loops, best of 11: 347 usec per loop
```