@tirkarthi 2019-02-16
Is `repeat(None, k)` with a for loop the fastest way to call a function k times in Python?