@rhettinger 2019-02-16
``repeat(None, k)`` is the fastest way to create an iterator that loops *k* times. 

Since the iterated action is doing more than just a function call, *map()* and *list()* won't be of any help in the loop body.